app: ChatActivity: show messages in user chats
User is now able to see message in chats, click on any "known" users card from MainActivity, and it'll open chatActivity.
ChatActivity is using FirebaseReyclerAdapter, which is used to handle UI stuff for an recyclerView (like other Adapters).
Create an Query object to get messages from :
	p2p_chats -> uniqueId

and attach the query is FirebaseRecyclerOptions with our Message class (model class for our message data structure )

* ChatActivity: Add FirebaseRecyclerAdapter to chatReyclerView, and use custom query
* ChatAdapter: Bind the data given by model class with single_message_view layout views, also change message card position according to sender receiver messages
* Message: Model class for Message object, which have senderUid, time and message
* Etc

Signed-off-by: adnan-44 <mohd.adnan0405@gmail.com>