app: ChatAdapter: Added functionality to update messageSeen field
Update the messageSeen field of message once the message is seen by receiver, to get this working, below is pseudo code for the same.
If currentUid and senderUid (message) are not equal, means receiver is gonna seen the message
so simply update the messageSeen field to true, to indicate that message is seen.
	Create and Map object, messageSeen(key), true (value)
	and use updateChildren() method on that message reference in firebase database

* ChatAdapter: Implemented above psedo code in onBind() method

Signed-off-by: adnan-44 <mohd.adnan0405@gmail.com>