app: MyProfileActivity: Allow users to update profile picture
By using Intent actions, get image which is needed to be uploaded, on successful the intent data will not be null, and if data from intent is not null then store that data in profileImageUri.
After that call UploadProfileImage() method with profileImageUri as parameter, inside this the image will be uploaded to Firebase storage
	profile_images -> Uid (where filename will be user's Uid)

And onSuccessfull uploading of image into firebase storage, it'll give downloadable Uri of newly updaloaded image, now call updateUserInfo() method with downloadable imageUri as parameter.
updateUserInfo() method will create an Map object with key (image_url) and value (downloadable imageUri in string), and will update the
	all_users_info -> Uid -> image_url (So, that database will have updated Uri to newly uploaded profile image)

* MyProfileActivity: get Image from user using intent actions, upload image using custom uploadProfileImage() method, and on successully uploading, call updateUserInfo() method to update user's profile info with update image_url.
* activity_my_profile layout: Added ImageView for profile image, and another imageView as (select image_ button
* gradle : Add dependencies for Firebase Storage
* Etc.

Signed-off-by: adnan-44 <mohd.adnan0405@gmail.com>