app: AddNewChatActivity: create methods to make code little clean
Wrap the user searching logics into custom methods, so that code may look little clean.
OnClick of search button, call searchUsers() method to find users in platform database. If that user is not found then show "User not found" toast, and if that user is found, our next tast is to check if that user is already known to current user or not.
To do so, alreadyKnown() method will be called, inside this method, it'll check whether usesr is already known to current user or not, If user is not "known" to current user, then show that "userFoundDialog" stuff (which will show user information in dialog box, with "cancel" and "Add" button".

* AddNewChatActivity: Code changes according to above given changes, split code in searchUsers() and alreadyKnown() methods, etc.
* Remove useless imports also

Signed-off-by: adnan-44 <mohd.adnan0405@gmail.com>