app: Only allow if user's email is verified
Previously, users were able to create account using any email addresses (even if that email doesn't exists). It was good, but when it comes to "password" recovery (or user forgotten their account password), it ended up in dead lock condition.
Because if "that" email doesn't exists, then we can't send password reset link to it. To overcome from this issue, users will have to "verify" their email addresses.
When user creating a new account, an "email" will be sent to their email address to "verify" account.

* CreateNewAccount : call sendEmailVerification() method to sent verificaton email, add "email_verified" field in users_info object.
* Manifest: Add new PendingAccoutActivity
* PendingAccount : A seperate activity for "post" account creation, it has refresh button to refresh current user's local firebase data.
  If user's email gets verified, then it'll automatically redirect user to the MainActivity.
* strings.xml : Added custom strings for "Pending account" activity
* etc.

Signed-off-by: adnan-44 <mohd.adnan0405@gmail.com>